2.6.32: Limit input-wacom to only provide "new" devices to RHEL 6.10
The driver included with RHEL 6.10 includes support for a few devices
that aren't supported by this input-wacom driver and vice-versa. For
example, RHEL supports the EKR but not the Cintiq Pro 32 while input-wacom
supports the 32 but not the EKR. This is a problem if you want to use
both the 32 and the EKR (or other combination of devices that aren't
both supported by one of the drivers).

As a workaround, we will have input-wacom install as a driver named
"wacom-ot" rather than "wacom" when a RHEL 6.10 system is detected.
This driver will *only* support devices that aren't already supported
by the existing RHEL driver.

This isn't ideal, but the proper fix (backporting EKR and others)
is more work than we're prepared to undertake at this point in time.

Signed-off-by: Ping Cheng <ping.cheng@wacom.com>
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>