Only use AM_EXTRA_RECURSIVE_TARGETS if defined
Commit 260ee967a0 added a call to the "AM_EXTRA_RECURSIVE_TARGETS"
macro in order to add "signature" as a recursive makefile target. This
allows users to run `make signature` from the top-level directory rather
than requring them to first change into the build directory for their
kernel. Unfortunately, "AM_EXTRA_RECURSIVE_TARGETS" is not available on
RHEL 6 since it requires Automake 1.13. This prevents our driver from
building from a git checkout on the OS (the tarball continues to work
fine).

To address this, we now only conditionally use the macro. If a RHEL 6
user builds from git (not a tarball), the `make signature` step will
not work unless they first change into the 2.6.32 directory.

Fixes: 260ee967a0 ("Add support for kernel module signing and enforcement")
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>