Improve tests to see if module signing is required
Previously we would rely on the 'mokutil' command to tell us if module
signing was required, but the check (seeing if secure boot was enabled)
isn't strictly correct and the utility may not exist anyway. This patch
has us read through the kernel config to see what module signing options
are enabled and reading the secure boot flag through a few interfaces
that may be available. This should give us a more correct view of if
module signing is actually required or not.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Acked-by: Ping Cheng <ping.cheng@wacom.com>