Error out if the build path has a space in it
The Linux kernel build system does not work properly if it is asked to
deal with spaces in paths. There's not much point making our own build
work in the presence of spaces, so at least for now make sure the autogen
script can run and that the generated configure script will print out an
error if this is attempted.

Ref: https://github.com/linuxwacom/input-wacom/issues/129
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>