configure: Force use of '-fno-pie' flag when test-compiling
Some compilers require the "-fno-pie" option to be passed when compiling
kernel code. If the flag is not present the compile will fail and may
cause our test-compile code to incorrectly take the failure path. In
particular, the "abs accessor" test may incorrectly identify a kernel
as needing the pre-2.6.36 API due to this issue.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>