Add workflows for automated testing via Github Actions
We've been using Travis for our automated tests, but that service is
becoming more difficult to use (open source projects have to periodically
request free credits to keep running). Github itself can be used as a
replacement by using its "Actions" infrastructure to execute various
defined workflows.

This commit ports over the Travis script to Github Actions with a few
minor modifications. In particular, the build environment is much newer
and was triggering several warnings which we've had to temporarily
mark with Wno-error. In addition we've needed to explicitly add "-fno-pie"
to the compile flags to make sure everything continues to compile.
The checkpatch action has been added as a seperate workflow since it
does not seem to be compatible with the "push" action.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>