Fix incorrect references to 'wacom' in Makefile for RHEL 7.4+ systems
We change the name of our module from "wacom" to "hid-wacom" on RHEL 7.4+
in order to follow along with the split-driver module which was adapted
for those systems. We apparently forgot to update several locations in
the code which refer to the "wacom" module by name, however.

This can prevent the hid-wacom driver from being loaded on RHEL 7.4
systems. This commit replaces the use of "wacom" with "$(MODULE_NAME)"
where needed. The Makefile should now properly sign the hid-wacom
driver and install the necessary depmod override.

Link: https://github.com/linuxwacom/input-wacom/issues/246
Fixes: 7aa0770a383c ("Use the kernel's built-in signing framework rather than our own")
Fixes: 102891af018e ("Add a special case for RHEL 7.4+")
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>