Add wacom_w8001 to the list of depmod overrides
Our depmod configuration dictates that the version of our modules found
in the "extra" modules subdirectory (e.g. /lib/modules/$(uname -r)/extra)
is to be preferred over any other version that may have been pre-installed
by the distro. We only specify the "wacom" module, however, which means
that its possible someone who specifically needs our out-of-tree version
of "wacom_w8001" wouldn't get it.

This commit adds the "wacom_w8001" driver to the list of overrides that
are searched for in the "extra" directory since we're installing it
alongside the "wacom" driver.

Link: https://github.com/linuxwacom/input-wacom/issues/246
Fixes: d4dbb98b2527 ("Prioritize compiled input-wacom modules with a depmod.d entry")
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>