Correct syntax error preventing RHEL/Fedora initrd from being updated
The Makefile syntax does not expect a colon at the end of the line. Its
presence here causes the test to fail, preventing dracut (the RHEL/Fedora
tool for updating initrd) from running. This is almost certianly the long-
standing source of issues with `make install` not appearing to have any
effect on the initrd.

Link: https://github.com/linuxwacom/input-wacom/issues/246
Fixes: b6e1bdaff833 ("Use `dracut` if available to update RHEL/Fedora initramfs")
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>