Configure dracut to install the input-wacom modules in the initramfs
The dracut tool can behave badly with the input-wacom driver when it
generates a new initramfs. In some cases it may prefer to install the
stock version of the driver to the initramfs rather than the input-wacom
version. When this occurs the user will experience different results
depending on if their tablet is connected at boot (and use the initramfs
driver) or connected after boot (and use the on-disk driver). It isn't
clear why this happens*, but it seems that we can work around the issue.

This commit adds a dracut configuration file which instructs the tool
to add our drivers to the install list. The code which handles this
list appears to reliably use the input-wacom version from the "extra"
directory rather than the stock version.

*The 90kernel-modules-extra dracut module finds the input-wacom version
of the driver and attempts to add it to the initramfs, but dracut is
configured to run in "hostonly" mode by 01-dist.conf. This mode causes
the installer to refuse to install the module if it isn't currently
loaded into the kernel. This isn't a problem on its own, but for some
reason a *different* part of the dracut process decides to ignore the
"hostonly" flag and install the stock version of the driver anyway.
Smells like a dracut bug...

Fixes: b6e1bdaff833 ("Use `dracut` if available to update RHEL/Fedora initramfs")
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>