github: Add 5.10 kernel to list of versions we test-compile against
The following imported commit depends on code from Linux 5.10, so we
should add that version to the list of kernels that we have Github's
infrastructure test-compile.

Note that this requires us to also switch the kernel config back to
using the frame-pointer based unwinder. Linux 4.15 made the new ORC
unwinder default, but that unwinder depends on some additional host
libraries and also fails to compile with the `-fno-pie` option that
we currently force.