3.7: Prevent segfault when connecting recent tablets
Several devices can trigger a segfault when they are plugged into a
system running a 3.7 kernel (e.g. RHEL 7.0 - 7.3). The failure is due
to `wacom_wac->shared->touch_input` being null when it is dereferenced
to check the device's PID in `wacom_setup_input_capabilities` at
wacom_wac.c:2707. Affected devices have type DTH1152T, DTH2452T,
WACOM_MSPROT, and WACOM_27QHDT.

Looking through the program flow, it quickly becomes clear that the
variable is only initialized toward the end of `wacom_probe` -- well
after the call to `wacom_setup_input_capabilities`. The variables used
by the initalization code appear to all be ready by the time that we
want to make this call, however, so we should be able to safely move
it earlier.

Fixes: af0ec069e12a ("backport: HID: wacom: support named keys on older devices")
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>