Build and install both 3.7 and 3.17 trees on RHEL 7.4+
RHEL 7.4 introduced a split-driver architecture which ships both a "wacom"
and "hid-wacom" module. The former is based on the driver that shipped with
Linux 3.10 (in the input subsystem) while the latter is based on a version
of the driver from 3.17 or later (in the hid subsystem).

We'd like to compile replacements for both modules so that users can use
input-wacom as a source of bugfixes for either. This commit appends to
the version variable rather than replacing it, which will trigger the
Makefile to start recursive builds in both directories. We also make sure
that both modules are marked to override the system versions in the depmod
configuration.

Link: https://github.com/linuxwacom/input-wacom/issues/273
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>