backport: HID: wacom: Correct power_supply type
POWER_SUPPLY_TYPE_USB seems to only ever be used by USB ports that are
used to charge the machine itself (so a "system" scope), like the
single USB port on a phone, rather than devices.

The wacom_sys driver is the only driver that sets its device battery as
being a USB type, which doesn't seem correct based on its usage, so
switch it to be a battery type like all the other USB-connected devices.

Signed-off-by: Bastien Nocera <hadess@hadess.net>
Acked-by: Jason Gerecke <jason.gerecke@wacom.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
[joshua.dickens@wacom.com: Imported into input-wacom repository (6ff1cae2e30a)]
Signed-off-by: Joshua Dickens <joshua.dickens@wacom.com>
[joshua.dickens@wacom.com: Backported from input-wacom repository (ed70da7a360e)]
Signed-off-by: Joshua Dickens <joshua.dickens@wacom.com>