github: Work around kernel build error
Older versions of the Linux kernel won't compile correctly on newer kernels.
This causes our automated patch checks to fail. This issue has been fixed
upstream but still needs to be applied to some older kernels.

This patch is specifically required to prevent a build failure in the 4.14
kernel that is added in the next commit.

Link: https://lore.kernel.org/selinux/20190225005528.28371-1-paulo@paulo.ac/
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>