3.17: Change conditional compilation of devm_add_action_or_reset
While this function was indeed added in v4.6 of the Linux kernel, it
has also been backported to RHEL 7.9. Because RHEL declares itself to
use a 3.10 kernel, however, the compile fails with the following
error about a redefined symbol:

./3.17/wacom_sys.c:66:12: error: redefinition of ‘devm_add_action_or_reset’
 static int devm_add_action_or_reset(struct device *dev,
            ^
In file included from include/linux/input.h:22:0,
                 from include/linux/hid.h:35,
                 from /home/mocaw/input-wacom/3.17/wacom_wac.h:7,
                 from /home/mocaw/input-wacom/3.17/wacom_sys.c:6:
include/linux/device.h:683:19: note: previous definition of ‘devm_add_action_or_reset’ was here
 static inline int devm_add_action_or_reset(struct device *dev,
                   ^

This commit adds a configure-time check for this function (like we
do for the power_supply APIs) so that code will compile properly
regardless of the declared kernel version.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>