configure.ac: Use MODULE_LICENSE when compiling test code
Attempts to build test code seem to all fail because MODULE_LICENSE
isn't defined. This can cause the build process to behave incorrectly
since it may wrongly believe a feature is not available when it really
is.

Adding MODULE_LICENSE("GPL") to the body of the test code causes the
configure-time checks to start working properly again. We can use the
"GPL" identifier here since our code is, indeed, GPL.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>