4.5: Change signature of wacom_i2c_probe depending on available API
Commit 0a3098db138c ("Input: wacom_i2c - Convert to i2c's .probe_new()")
upstream updated the driver to make use of a modernized API that was
introduced in Linux 4.10. The "probe_new" function that was introduced
to make the API migration easier was eventually renamed to just "probe"
in Linux 6.3 -- breaking out-of-tree drivers that had not yet updated.

This commit checks to see if the modern version of the API is required
or not. If it is, we modify the function signature to allow it to work
properly.

Fixes: https://github.com/linuxwacom/input-wacom/issues/367
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>