github: Install libelf-dev kernel compile dependency and fixup PIE
The Linux kernel has had a somewhat-flexible dependency on headers
provided by the libelf-dev package for a while. As long as we compiled
with the options `-fno-pie -no-pie`, however, the build would succeed.
Unfortunately, newer kernels don't like it when we manually specify
those options. They now complain that some parts of the code *must*
be compiled with PIE... Before we're able to add these newer kernels
to the test matrix, we must get ourselves out of this jam...

The solution that this commit implements is to only set those options
on the older kernels that actually require it. Around the time of Linux
4.9, the kernel updated its own makefile to add these options when
required. Only setting these flags when required prevents the problems
from cropping up.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>