Makefile.in: Allow extra make flags to be passed to the kernel
It can be useful to pass additional flags to sub-make run in the kernel's
source tree. For example, you might want to adjust the warning levels or
modify some behavior of the kernel's Makefiles. To enable this ability,
when input-wacom is built with an EXTRA_MAKEFLAGS variable set, its
contents will be passed to the sub-make.

For example, to enable additional kernel warnings, you can now call the
input-wacom make as follows: `$ make W=123`

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>