Update checkpatch to run only when kernel files are touched
Updating the checkpatch github action so that it only runs when we
modify kernel related files (ex: Anything in 3.17/4.5 sub directories).

This change keeps the tests run during a PR cleaner by only running
when it is most relevant for code styling/commit messages that are
relevant to keeping style parity with upstream, while ignoring
irrelevant changes to files that aren't upstream (such as this one).