github: Fix incorrect 'version' file generation
The 'version' file is created as part of the release process and is
supposed to contain the version number of the tarball. The contents
of this file become part of /sys/module/wacom/version, which we rely
on to understand what driver version users are running.

Recent release (1.0.0, 1.1.0, and 1.2.0) have had faulty version files
which just contain the text "input-wacom". The issue was traced to
the checkout action used by our tagged-release workflow. This action
does not include tags by default, which prevents the git-version-gen
script from being able to discover what version of the code is being
built. In the absence of a git tag, it falls back to using the directory
name instead: 'input-wacom'.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>