configure.ac: Ingore modpost errors from WACOM_LINUX_COMPILE_IFELSE
The WACOM_LINUX_COMPILE_IFELSE macro performs test builds of small code
snippets. It executes different sections of the configure script depending
on if the build failed or succeeded. We use this function to test whether
various kernel features are present or not. The macro performs a full
`make modules` on these code snippets, which includes having "modpost"
link the object files into .ko files.

This final linking step is neither necessary nor desired. We're only
interested in if the code snippets /compile/ and don't care if they
link. Trying to link them can cause the macro to incorrectly take the
"fail" even when the compile itself succeeds. This is mostly a problem
on github runners -- see commit 9b6c25523407 ("github: Do not fail
with error when modpost can't find symbols") for the same issue but
with the overall build instead.

This commit adds `KBUILD_MODPOST_WARN=1` to the test build commands
to transform link errors into harmless warnings.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>

Fix: Nowarn